atualização das classes para vo com mapper

+++ src/main/java/br/com/udidrive/vo/FotoVO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class FotoVO {
     private Long idFoto;
-    private Integer entidadeTipo; // ID do tipo de entidade (cliente, motorista, veículo)
-    private Integer entidadeId;   // ID da entidade associada (cliente, motorista ou veículo)
+    private Long entidadeTipo; // ID do tipo de entidade (cliente, motorista, veículo)
+    private Long entidadeId;   // ID da entidade associada (cliente, motorista ou veículo)
     private byte[] foto;          // Foto em formato binário
     private String criadoEm;      // Data de criação, formatada para transferência
 }
